backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
